Separate and reorganise discovery pass
