+++ compiler/hash-typecheck/src/new/passes/resolution/paths.rs
